+++ packages/client/hmi-client/src/services/calibrate-workflow.ts
@@ -1,23 +1,12 @@
 import * as d3 from 'd3';
 
-import type { ModelConfiguration, Dataset, CsvAsset, State, DatasetColumn } from '@/types/Types';
+import type { ModelConfiguration, Dataset, CsvAsset, State } from '@/types/Types';
 import { getModelConfigurationById } from '@/services/model-configurations';
 import { downloadRawFile, getDataset } from '@/services/dataset';
-import { getEntitySimilarity } from './concept';
 
 export interface CalibrateMap {
 	modelVariable: string;
 	datasetVariable: string;
-}
-
-interface Entity {
-	id: string;
-	groundings?: any[];
-}
-
-interface EntityMap {
-	source: string;
-	target: string;
 }
 
 // Used in the setup of calibration node and drill down
@@ -125,156 +114,4 @@ export const renderLossGraph = (
 		yAxisGroup = svg.append('g').attr('class', 'y-axis');
 	}
 	yAxisGroup.attr('transform', `translate(${marginLeft}, 0)`).call(yAxis);
-};
-
-// Takes in 2 lists of generic {id, groundings} and returns the singular
-// closest match for each element in list one
-export const autoEntityMapping = async (
-	sourceEntities: Entity[],
-	targetEntities: Entity[],
-	acceptableDist?: number
-) => {
-	const result = [] as EntityMap[];
-	const acceptableDistance = acceptableDist ?? 0.5;
-
-	// join all source and target for 1 mira call
-	const allSourceGroundings = sourceEntities.flatMap(({ groundings }) => groundings ?? []);
-	const allTargetGroundings = targetEntities.flatMap(({ groundings }) => groundings ?? []);
-
-	// Take out any duplicates
-	const distinctSourceGroundings = [...new Set(allSourceGroundings)];
-	const distinctTargetGroundings = [...new Set(allTargetGroundings)];
-	const allSimilarity = await getEntitySimilarity(
-		distinctSourceGroundings,
-		distinctTargetGroundings
-	);
-	if (!allSimilarity) return result;
-	// Filter out anything with a similarity too small
-	const filteredSimilarity = allSimilarity.filter((ele) => ele.similarity >= acceptableDistance);
-	const validSources: any[] = [];
-	const validTargets: any[] = [];
-
-	// Join results back with source and target
-	filteredSimilarity.forEach((similarity) => {
-		// Find all Sources assosiated with this similarity
-		sourceEntities.forEach((source) => {
-			if (source.groundings?.includes(similarity.source)) {
-				validSources.push({
-					sourceId: source.id,
-					sourceKey: similarity.source,
-					targetKey: similarity.target,
-					distance: similarity.similarity
-				});
-			}
-		});
-		// Find all targets assosiated with this similarity
-		targetEntities.forEach((target) => {
-			if (target.groundings?.includes(similarity.target)) {
-				validTargets.push({
-					targetId: target.id,
-					sourceKey: similarity.source,
-					targetKey: similarity.target,
-					distance: similarity.similarity
-				});
-			}
-		});
-	});
-
-	// for each distinct source, find its highest matching target:
-	const distinctSource = [...new Set(validSources.map((ele) => ele.sourceId))];
-	distinctSource.forEach((distinctSourceId) => {
-		let currentDistance = -Infinity;
-		let currentTargetId = '';
-		validSources.forEach((source) => {
-			if (distinctSourceId === source.sourceId) {
-				validTargets.forEach((target) => {
-					// Note here we are using the source and target groundings as a key
-					if (
-						source.sourceKey === target.sourceKey &&
-						source.targetKey === target.targetKey &&
-						target.distance > currentDistance
-					) {
-						currentDistance = target.distance;
-						currentTargetId = target.targetId;
-					}
-				});
-			}
-		});
-		result.push({ source: distinctSourceId, target: currentTargetId });
-	});
-
-	return result;
-};
-
-// Takes in a list of states and a list of datasetcolumns.
-// Transforms them into generic entities with {id, groundings}
-// Uses autoEntityMapping to determine 1:1 mapping
-// rewrites result in form {modelVariable, datasetVariable}
-export const autoCalibrationMapping = async (
-	modelOptions: State[],
-	datasetOptions: DatasetColumn[]
-) => {
-	const result = [] as CalibrateMap[];
-	const acceptableDistance = 0.5;
-	const sourceEntities: Entity[] = [];
-	const targetEntities: Entity[] = [];
-
-	// Fill sourceEntities with modelOptions
-	modelOptions.forEach((state) => {
-		const groundings = state.grounding?.identifiers
-			? Object.entries(state.grounding.identifiers).map((ele) => ele.join(':'))
-			: undefined;
-		sourceEntities.push({ id: state.id, groundings });
-	});
-
-	// Fill targetEntities with datasetOptions
-	datasetOptions.forEach((col) => {
-		const groundings = col.metadata?.groundings?.identifiers
-			? Object.entries(col.metadata.groundings.identifiers).map((ele) => ele.join(':'))
-			: undefined;
-		targetEntities.push({ id: col.name, groundings });
-	});
-
-	const entityResult = await autoEntityMapping(sourceEntities, targetEntities, acceptableDistance);
-
-	// rename result to CalibrateMap for users of this function
-	entityResult.forEach((entity) => {
-		result.push({ modelVariable: entity.source, datasetVariable: entity.target });
-	});
-	return result;
-};
-
-// Takes in two lists of states
-// Transforms them into generic entities with {id, groundings}
-// Uses autoEntityMapping to determine 1:1 mapping
-// rewrites result in form {modelOneVariable, modelTwoVariable}
-export const autoModelMapping = async (modelOneOptions: State[], modelTwoOptions: State[]) => {
-	const result = [] as any[];
-	const acceptableDistance = 0.5;
-	const sourceEntities: Entity[] = [];
-	const targetEntities: Entity[] = [];
-
-	// Fill sourceEntities with modelOneOptions
-	modelOneOptions.forEach((state) => {
-		const groundings = state.grounding?.identifiers
-			? Object.entries(state.grounding.identifiers).map((ele) => ele.join(':'))
-			: undefined;
-		sourceEntities.push({ id: state.id, groundings });
-	});
-
-	// Fill targetEntities with datasetOptions
-	modelTwoOptions.forEach((state) => {
-		const groundings = state.grounding?.identifiers
-			? Object.entries(state.grounding.identifiers).map((ele) => ele.join(':'))
-			: undefined;
-		targetEntities.push({ id: state.id, groundings });
-	});
-
-	const entityResult = await autoEntityMapping(sourceEntities, targetEntities, acceptableDistance);
-
-	// rename result to CalibrateMap for users of this function
-	entityResult.forEach((entity) => {
-		result.push({ modelOneVariable: entity.source, modelTwoVariable: entity.target });
-	});
-	return result;
 };
